Move the 'BalanceManager' to its own package
